# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Payment Orchestration API
  description: |
    Payment authorization orchestration platform responsible for PSP routing,
    retry handling, idempotency enforcement, reconciliation, provider failover,
    observability, and asynchronous recovery.

    **Scope**: Authorization lifecycle only. Settlement, ledgering, capture,
    refund, and acquiring workflows are explicitly out of scope.

    ## Request Headers

    | Header | Required | Behavior |
    |---|---|---|
    | `X-Request-Id` | YES | Rejected with `400` if absent or malformed. Echoed in all responses as `request_id`. |
    | `X-Correlation-Id` | OPTIONAL | System generates a UUID and returns it if absent. Propagated across all internal service boundaries and downstream PSP calls. |
    | `internal_request_id` | SYSTEM GENERATED | Always generated internally. Never exposed in API responses. Used solely for internal observability, log correlation, and tracing pipelines. |

    ## Authentication & Key Rotation

    Every request must be signed using HMAC-SHA256. The canonical string is:
    ```
    HTTP_METHOD + REQUEST_PATH + BODY_SHA256 + TIMESTAMP + NONCE + MERCHANT_ID
    ```
    Signature: `Base64(HMAC_SHA256(merchant_secret, canonical_string))`

    Merchant secrets are stored in a **secrets manager / KMS**, never in application config.
    Keys are versioned (e.g. `merchant_key_v1`, `merchant_key_v2`). During a rotation window
...
    expires the old key is invalidated.

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/payments-orchestration/webhooks/{provider}:
    post:
      tags: [Webhooks]
      summary: Receive PSP Webhook
      description: |
        Ingests asynchronous callbacks from PSP providers. Used to:
        - Resolve `PENDING` payments to final states (`AUTHORIZED` or `FAILED`).
        - Confirm authorizations not returned in the synchronous flow.
        - Accelerate reconciliation.

        **Security:** Each provider webhook is verified using the provider-specific
        signature scheme. Unverified webhooks are rejected with `401`.

        **Deduplication:** Webhook events are deduplicated by `event_id`.
        Replayed events are acknowledged with `200` but not reprocessed.

        **Idempotency:** Processing is replay-safe. Multiple deliveries of the
        same event produce the same outcome.
      operationId: receiveWebhook
      parameters:
        - name: provider
          in: path
          required: true
          description: PSP provider identifier.
          schema:
            type: string
            enum: [PSP_A, PSP_B]
            example: "PSP_A"
        - name: X-PSP-Signature
          in: header
          required: true
          description: |
            Provider-specific webhook signature for payload verification.
            Format varies by provider (HMAC-SHA256, RSA, etc.).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
            examples:
              authorized:
                summary: PSP confirms authorization
                value:
                  event_id: "evt_psp_abc123"
                  event_type: "PAYMENT_AUTHORIZED"
                  provider_reference: "provider_tx_9f8e7d6c"
                  intent_id: "7f3e2c1a-4b5d-6e7f-8a9b-0c1d2e3f4a5b"
                  attempt_id: "1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d"
                  status: "AUTHORIZED"
                  timestamp: "2026-05-27T12:00:05Z"
                  metadata:
                    authorization_code: "AUTH789012"
                    network: "VISA"
              failed:
                summary: PSP confirms definitive failure
                value:
                  event_id: "evt_psp_abc124"
                  event_type: "PAYMENT_FAILED"
                  provider_reference: "provider_tx_9f8e7d6d"
                  intent_id: "7f3e2c1a-4b5d-6e7f-8a9b-0c1d2e3f0000"
                  attempt_id: "1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b0000"
                  status: "FAILED"
                  timestamp: "2026-05-27T12:01:00Z"
                  metadata:
                    decline_code: "INSUFFICIENT_FUNDS"
                    network: "MASTERCARD"
      responses:
        '200':
          description: Webhook acknowledged and processed (or already processed — idempotent).
          content:
            application/json:
              schema:
                type: object
                properties:
                  acknowledged:
                    type: boolean
                    example: true
                  event_id:
                    type: string
                    example: "evt_psp_abc123"

        '400':
          description: Malformed webhook payload — missing required fields or invalid structure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "VALIDATION_ERROR"
                message: "Webhook payload is missing required field 'event_id'"
                details:
                  - field: "event_id"
                    issue: "must not be null"
                request_id: "req_wh001"
                timestamp: "2026-05-27T12:00:00Z"

        '401':
          description: |
            PSP webhook signature verification failed. Payload may have been
            tampered with, or the signature secret is misconfigured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "INVALID_WEBHOOK_SIGNATURE"
                message: "Webhook signature verification failed for provider PSP_A"
                details: []
                request_id: "req_wh002"
                timestamp: "2026-05-27T12:00:00Z"

        '404':
          description: No matching payment intent or attempt found for the webhook's references.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "PAYMENT_NOT_FOUND"
                message: "No payment attempt found for provider_reference 'provider_tx_unknown'"
                details: []
                request_id: "req_wh003"
                timestamp: "2026-05-27T12:00:00Z"

        '422':
          description: |
            Webhook event is structurally valid but cannot be applied.
            Examples: illegal state transition, or event references a superseded attempt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                illegal_transition:
                  summary: Illegal state transition
                  value:
                    error_code: "ILLEGAL_STATE_TRANSITION"
                    message: "Cannot transition payment from FAILED to AUTHORIZED"
                    details:
                      - field: "status"
                        issue: "FAILED is a terminal state"
                    request_id: "req_wh004"
                    timestamp: "2026-05-27T12:00:00Z"
                superseded_attempt:
                  summary: Attempt already superseded
                  value:
                    error_code: "ATTEMPT_SUPERSEDED"
                    message: "The referenced attempt has been superseded by a successful retry"
                    details: []
                    request_id: "req_wh005"
                    timestamp: "2026-05-27T12:00:00Z"

        '500':
          $ref: '#/components/responses/InternalError'

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────
components:
...
  schemas:
...
    WebhookRequest:
      type: object
      required:
        - event_id
        - event_type
        - provider_reference
        - status
        - timestamp
      properties:
        event_id:
          type: string
          description: PSP-generated unique event ID. Used for deduplication.
          example: "evt_psp_abc123"
        event_type:
          type: string
          enum: [PAYMENT_AUTHORIZED, PAYMENT_FAILED, PAYMENT_PENDING]
          example: "PAYMENT_AUTHORIZED"
        provider_reference:
          type: string
          description: PSP transaction reference. Must match a known payment_attempts.provider_reference.
          example: "provider_tx_9f8e7d6c"
        intent_id:
          type: string
          format: uuid
          nullable: true
          description: |
            Optional. If provided, used to correlate directly to a payment intent.
            If absent, correlation is done via provider_reference.
          example: "7f3e2c1a-4b5d-6e7f-8a9b-0c1d2e3f4a5b"
        attempt_id:
          type: string
          format: uuid
          nullable: true
          description: Optional. If provided, correlates directly to a specific attempt.
          example: "1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d"
        status:
          type: string
          enum: [AUTHORIZED, FAILED, PENDING]
          example: "AUTHORIZED"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when the event occurred at the PSP.
          example: "2026-05-27T12:00:05Z"
        metadata:
          type: object
          description: Provider-specific event metadata (authorization codes, decline codes, etc.).
          additionalProperties: true
          example:
            authorization_code: "AUTH789012"
            network: "VISA"
    ValidationDetail:
      type: object
      properties:
        field:
          type: string
          description: JSON path or header name of the offending field.
          example: "transaction_amount.currency_code"
        issue:
          type: string
          description: Human-readable description of the constraint violation.
          example: "must be a valid ISO 4217 3-letter currency code"
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
        - details
        - request_id
        - timestamp
      properties:
        error_code:
          type: string
          description: Machine-readable error code for programmatic handling.
          enum:
            - VALIDATION_ERROR
            - MISSING_REQUEST_ID
            - INVALID_SIGNATURE
            - NONCE_REUSED
            - TIMESTAMP_INVALID
            - IDEMPOTENCY_CONFLICT
            - MERCHANT_INACTIVE
            - DUPLICATE_ORDER_ID
            - UNSUPPORTED_CURRENCY
            - UNSUPPORTED_PAYMENT_METHOD
            - INVALID_AMOUNT
            - SETTLEMENT_INCOMPATIBLE
            - PAYMENT_NOT_FOUND
            - ILLEGAL_STATE_TRANSITION
            - ATTEMPT_SUPERSEDED
            - INVALID_WEBHOOK_SIGNATURE
            - INTERNAL_ERROR
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error description suitable for logging.
          example: "Validation failed"
        details:
          type: array
          description: Field-level validation details. Empty array for non-validation errors.
          items:
            $ref: '#/components/schemas/ValidationDetail'
        request_id:
          type: string
          description: |
            Echo of the X-Request-Id header. If X-Request-Id was absent or malformed,
            this contains a system-generated trace ID prefixed with sys_ to distinguish
            it from client-provided IDs. Always present.
          example: "req_abc123"
        timestamp:
          type: string
          format: date-time
          example: "2026-05-27T12:00:00Z"
  responses:
...
    InternalError:
      description: Unexpected internal system failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "INTERNAL_ERROR"
            message: "An unexpected error occurred. Please retry or contact support."
            details: []
            request_id: "req_000"
            timestamp: "2026-05-27T12:00:00Z"